+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed [3.5.25] - 21-07-2026
+
+- Only show the revocation button for credentials whose credential status is of type `BitstringStatusListEntry`. This hides the button for legacy credentials that have `PlainListEntity` credential status.
+
 ## [3.5.24] - 21-07-2026
 
 ### Fixed